:boom: Refactor Updater function to return error
Removed dependency on jog package and revised the Updater function to return an error. The jog.Error log call has been replaced with an actual error return. Created a new function named MustUpdater that will be used when an error return is not needed. Updated unit tests to reflect these changes.